requirements.txt
streamlit==1.17.0
psycopg2-binary==2.9.5